try-domain-aw/requirements.txt
streamlit==0.90.0
mysql-connector-python==8.0.28